Restructure plugin to support both PocketMine and CLI access
This commit lays down a basic framework to support both PocketMine and CLI-based interfacing. It abstracts common functionality away from the PocketMine API (such as async operations) into interfaces and implements them with CLI counterparts.

In addition to restructuring the plugin for multiple interfaces, this commit also replaces the old callback-based asynchronous model with a Promise-based model with the react/promise package in composer. This allows arguably much prettier asynchronous programming and made the abstraction job a little easier.